Fix: 파일 및 폴더 구조 변경
