+++ client/src/App.tsx
@@ -1,18 +1,18 @@
 import { Routes, Route } from 'react-router-dom';
 import loadable from '@loadable/component';
-import Login from './Routes/Login';
-import Signup from './Routes/Signup';
+import Login from './Routes/Login/index';
+import Signup from './Routes/Signup/index';
 
-const Home = loadable(() => import('Routes/Home'));
+const Home = loadable(() => import('Routes/Home/index'));
 const Profile = loadable(() => import('Routes/Profile'));
 const Mypost = loadable(() => import('Routes/Mypost'));
 const Mycomment = loadable(() => import('Routes/Mycomment'));
-const Search = loadable(() => import('Routes/Search'));
+const Search = loadable(() => import('Routes/Search/index'));
 const Post = loadable(() => import('Routes/Post'));
 const PostView = loadable(() => import('Routes/PostView'));
 const PostModify = loadable(() => import('Routes/PostModify'));
 const Comment = loadable(() => import('Routes/Comment'));
-const Chat = loadable(() => import('Routes/Chat'));
+const Chat = loadable(() => import('Routes/Chat/index'));
 const Naver = loadable(() => import('Routes/Auth/Naver'));
 const Kakao = loadable(() => import('Routes/Auth/Kakao'));
 const Google = loadable(() => import('Routes/Auth/Google'));
