+++ client/src/Routes/Login/styles.tsx
@@ -1,18 +1,9 @@
 import styled from 'styled-components';
-import loadable from '@loadable/component';
-import { useState, useCallback } from 'react';
-import axios from 'axios';
-import { useDispatch } from 'react-redux';
-import { useNavigate } from 'react-router';
-import { AppDispatch } from 'index';
-import { UserLogin } from '../index';
-import kakao from '../Image/Btn/kakao.png';
-import google from '../Image/Btn/google.png';
-import naver from '../Image/Btn/naver.png';
+import kakao from '../../Image/Btn/kakao.png';
+import google from '../../Image/Btn/google.png';
+import naver from '../../Image/Btn/naver.png';
 
-const Home = loadable(() => import('Routes/Home'));
-
-const LoginWrapper = styled.div`
+export const LoginWrapper = styled.div`
   width: 100%;
   height: 100%;
   position: absolute;
@@ -24,7 +15,7 @@ const LoginWrapper = styled.div`
     top: 20px;
   }
 `;
-const Wrapper = styled.div`
+export const Wrapper = styled.div`
   width: 370px;
   height: 480px;
   display: flex;
@@ -44,13 +35,13 @@ const Wrapper = styled.div`
     width: 270px;
   }
 `;
-const Form = styled.form`
+export const Form = styled.form`
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
 `;
-const EmailInput = styled.input`
+export const EmailInput = styled.input`
   width: 315px;
   font-size: ${(props) => props.theme.fontSize.small};
   color: ${(props) => props.theme.lightGrey};
@@ -65,7 +56,7 @@ const EmailInput = styled.input`
     width: 250px;
   }
 `;
-const PasswordInput = styled.input`
+export const PasswordInput = styled.input`
   width: 315px;
   font-size: ${(props) => props.theme.fontSize.small};
   font-weight: normal;
@@ -81,7 +72,7 @@ const PasswordInput = styled.input`
     width: 250px;
   }
 `;
-const ExitBtn = styled.button`
+export const ExitBtn = styled.button`
   color: ${(props) => props.theme.lightGrey};
   font-size: ${(props) => props.theme.fontSize.small};
   background-color: inherit;
@@ -91,7 +82,7 @@ const ExitBtn = styled.button`
   right: 25px;
   cursor: pointer;
 `;
-const LoginBtn = styled.button`
+export const LoginBtn = styled.button`
   width: 240px;
   background-color: ${(props) => props.theme.lightGrey};
   color: ${(props) => props.theme.white};
@@ -114,7 +105,7 @@ const LoginBtn = styled.button`
     width: 180px;
   }
 `;
-const SignupBtn = styled.div`
+export const SignupBtn = styled.div`
   width: 220px;
   background-color: ${(props) => props.theme.white};
   color: ${(props) => props.theme.lightGrey};
@@ -139,7 +130,7 @@ const SignupBtn = styled.div`
   }
 `;
 
-const Naver = styled.a`
+export const Naver = styled.a`
   width: 225px;
   height: 48px;
   display: flex;
@@ -161,7 +152,7 @@ const Naver = styled.a`
   }
 `;
 
-const Kaoko = styled(Naver)`
+export const Kaoko = styled(Naver)`
   background-image: url(${kakao});
   @media ${(props) => props.theme.iPhone12Pro} {
     margin-bottom: 0px;
@@ -173,7 +164,7 @@ const Kaoko = styled(Naver)`
   }
 `;
 
-const Google = styled(Naver)`
+export const Google = styled(Naver)`
   background-image: url(${google});
   width: 230px;
   border-radius: 10px;
@@ -190,7 +181,7 @@ const Google = styled(Naver)`
   }
 `;
 
-const Text = styled.div`
+export const Text = styled.div`
   color: ${(props) => props.theme.lightGrey};
   font-size: ${(props) => props.theme.fontSize.tiny};
   margin-bottom: 15px;
@@ -201,105 +192,3 @@ const Text = styled.div`
     font-size: ${(props) => props.theme.fontSize.micro};
   }
 `;
-
-function Login() {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [errorMessage, setErrorMessage] = useState('');
-  const [open, isOpen] = useState(true);
-  const navigate = useNavigate();
-  const dispatch = useDispatch<AppDispatch>();
-
-  const handleOnSubmit = useCallback(
-    async (e: React.FormEvent<HTMLFormElement>) => {
-      e.preventDefault();
-      try {
-        if (email && password) {
-          const data = await axios.post(
-            `${process.env.REACT_APP_SERVER}/users/login`,
-            { email, password },
-            {
-              headers: {
-                'Content-Type': 'application/json',
-              },
-              withCredentials: true,
-            },
-          );
-          dispatch(UserLogin(data.data.data));
-          setEmail('');
-          setPassword('');
-          navigate('/');
-        }
-      } catch (err) {
-        setErrorMessage('이메일과 비밀번호를 다시 확인해주세요.');
-      }
-    },
-    [email, password, setEmail, setPassword, navigate, setErrorMessage],
-  );
-
-  const emailOnChange = useCallback(
-    (e: React.FormEvent<HTMLInputElement>) => {
-      setEmail(e.currentTarget.value);
-    },
-    [setEmail],
-  );
-
-  const passwordOnChange = useCallback(
-    (e: React.FormEvent<HTMLInputElement>) => {
-      setPassword(e.currentTarget.value);
-    },
-    [setPassword],
-  );
-
-  const exitOnClick = useCallback(() => {
-    isOpen(false);
-    navigate('/');
-  }, [open, isOpen, navigate]);
-
-  const signUpOnClick = useCallback(() => {
-    navigate('/signup');
-  }, [navigate]);
-
-  const state = window.localStorage.getItem('com.naver.nid.oauth.state_token');
-  const naverUrl = `https://nid.naver.com/oauth2.0/authorize?response_type=code&client_id=${process.env.REACT_APP_NAVER_CLIENT_ID}&state=${state}&redirect_uri=${process.env.REACT_APP_NAVER_CALLBACK_URL}`;
-  const kakaoUrl = process.env.REACT_APP_KAKAO_AUTH_URL;
-  const googleUrl = process.env.REACT_APP_GOOGLE_AUTH_URL;
-
-  return (
-    <>
-      {open ? (
-        <LoginWrapper>
-          <Wrapper>
-            <ExitBtn onClick={exitOnClick}>X</ExitBtn>
-            <Form onSubmit={handleOnSubmit}>
-              <EmailInput
-                value={email}
-                required
-                onChange={emailOnChange}
-                placeholder="이메일을 입력해주세요"
-              />
-              <PasswordInput
-                value={password}
-                required
-                type="password"
-                onChange={passwordOnChange}
-                placeholder="비밀번호을 입력해주세요"
-              />
-              <Text>{errorMessage || null}</Text>
-              <LoginBtn type="submit">로그인</LoginBtn>
-            </Form>
-            <Text>------------------------ 또는 -----------------------</Text>
-            <Naver href={naverUrl} />
-            <Kaoko href={kakaoUrl} />
-            <Google href={googleUrl} />
-            <Text>--------------- 아직 회원이 아니시라면? --------------</Text>
-            <SignupBtn onClick={signUpOnClick}>회원가입</SignupBtn>
-          </Wrapper>
-        </LoginWrapper>
-      ) : null}
-      <Home />
-    </>
-  );
-}
-
-export default Login;
